+++ lib/src/controller/osm_controller.dart
@@ -34,16 +34,21 @@ class OSMController {
   }
 
   /// initMap: initialisation of osm map
-  /// [initPosition] : (geoPoint) animate map to initPosition
-  /// [initWithUserPosition] : set map in user position
+  /// [initPosition]          : (geoPoint) animate map to initPosition
+  /// [initWithUserPosition]  : set map in user position
+  /// [box]                   : (BoundingBox) area limit of the map
   Future<void> initMap({
     GeoPoint? initPosition,
     bool initWithUserPosition = false,
+    BoundingBox? box,
   }) async {
     /// load config map scene for iOS
     if (Platform.isIOS) {
       await osmPlatform.initIosMap(_idMap);
     }
+
+
+    _checkBoundingBox(box, initPosition);
 
     osmPlatform.setDefaultZoom(_idMap, _osmFlutterState.widget.defaultZoom);
 
@@ -80,6 +85,9 @@ class OSMController {
           print(err);
         });*/
     }
+    if (Platform.isIOS) {
+      await osmPlatform.initIosMap(_idMap);
+    }
 
     /// change default icon  marker
     final defaultIcon = _osmFlutterState.widget.markerOption?.defaultMarker;
@@ -108,7 +116,12 @@ class OSMController {
     /// init location in map
     if (initWithUserPosition && !_osmFlutterState.widget.isPicker) {
       initPosition = await myLocation();
+      _checkBoundingBox(box, initPosition);
     }
+    if (box != null && !box.isWorld()) {
+      await limitAreaMap(box);
+    }
+
     if (initPosition != null) {
       await osmPlatform.initMap(
         _idMap,
@@ -160,6 +173,15 @@ class OSMController {
     }
   }
 
+  void _checkBoundingBox(BoundingBox? box, GeoPoint? initPosition) {
+    if (box != null && !box.isWorld() && initPosition != null) {
+      if (!box.inBoundingBox(initPosition)) {
+        throw Exception(
+            "you want to limit the area of the map but your init location is already outside the area!");
+      }
+    }
+  }
+
   Future _initializeRoadInformation() async {
     await osmPlatform.setColorRoad(
       _idMap,
@@ -172,6 +194,21 @@ class OSMController {
         _osmFlutterState.endIconKey,
         _osmFlutterState.middleIconKey
       ],
+    );
+  }
+
+  /// set area camera limit of the map
+  /// [box] : (BoundingBox) bounding that map cannot exceed from it
+  Future<void> limitAreaMap(BoundingBox box) async {
+    await osmPlatform.limitArea(
+      _idMap,
+      box,
+    );
+  }
+  /// remove area camera limit from the map
+  Future<void> removeLimitAreaMap() async {
+    await osmPlatform.removeLimitArea(
+      _idMap,
     );
   }